requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.3
click==8.0.1
ConfigArgParse==1.5.1
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.5
django-rest-swagger==2.2.0
djangorestframework==3.12.4
drf-yasg==1.20.0
Flask==2.0.1
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
gevent==21.1.2
geventhttpclient==1.4.4
greenlet==1.1.0
idna==3.2
inflection==0.5.1
itsdangerous==2.0.1
itypes==1.2.0
Jinja2==3.0.1
locust==2.0.0
MarkupSafe==2.0.1
msgpack==1.0.2
mysqlclient==2.0.3
openapi-codec==1.3.2
packaging==21.0
psutil==5.8.0
pyparsing==2.4.7
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.1.0
requests==2.26.0
roundrobin==0.0.2
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
simplejson==3.17.3
six==1.16.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.6
Werkzeug==2.0.1
zope.event==4.5.0
zope.interface==5.4.0
